+++ src/main/java/com/insane96mcp/shieldsplus/mixin/LivingEntityMixin.java
@@ -9,7 +9,7 @@ import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.Constant;
 import org.spongepowered.asm.mixin.injection.ModifyConstant;
 
-@Mixin(LivingEntity.class)
+@Mixin(value = LivingEntity.class, priority = 1001)
 public abstract class LivingEntityMixin extends Entity {
 
     public LivingEntityMixin(EntityType<?> p_19870_, Level p_19871_) {