+++ src/main/java/com/insane96mcp/shieldsplus/setup/Strings.java
@@ -1,5 +1,7 @@
 package com.insane96mcp.shieldsplus.setup;
 
+import com.insane96mcp.shieldsplus.ShieldsPlus;
+
 public class Strings {
     public static class Items {
         public static final String WOODEN_SHIELD = "wooden_shield";
@@ -14,6 +16,6 @@ public class Strings {
     }
 
     public static class Translatable {
-
+        public static final String DAMAGE_BLOCKED = ShieldsPlus.RESOURCE_PREFIX + "damage_blocked";
     }
 }