+++ changelog.md
@@ -1,4 +1,8 @@
 # Changelog
 
+## 1.0.1
+* Added "blocking damage" tooltip on Shields
+* Raised mixin priority, should prevent IguanaTweaks from messing with the blocking delay (0.25 secs)
+
 ## 1.0.0
 * First Release with shields and a few small features